Refactor presentation component organization and add network service state guards
Organize UI components by moving `Sync360Surface` from the `brandComponents` package to a new `presentationComponents` package and updating relevant imports in `ReceiveScreen` and `SendScreen`.

Improve the robustness of `AndroidNetworkServices` by adding state-based guards to lifecycle methods:
- `startNetworkServices` now ignores requests if the discovery service is not in an `Idle` state.
- `stopDiscoveryServices` only proceeds if the service is currently `Running`.
- `restartDiscoveryServices` is restricted to when the service is `Idle`.

These changes prevent redundant service operations and potential crashes resulting from invalid state transitions during peer discovery.
